Install requirements only if file exists.

+++ ci/docker/install/centos7_python.sh
@@ -34,5 +34,6 @@ make install
 cd ..
 rm -rf Python-$PYTHON_VERSION*
 pip3 install --upgrade pip setuptools wheel
-pip3 install -r /work/requirements
-
+if [[ -f /work/requirements ]]; then
+    pip3 install -r /work/requirements
+fi
